requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
google-auth==2.42.1
    # via google-genai
google-genai==1.47.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via google-genai
pydantic-core==2.41.4
    # via pydantic
requests==2.32.5
    # via google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via google-genai
